jpa annotation 추가
