--- server/src/main/java/domain/Insurance.java
@@ -1,15 +1,0 @@
-package domain;
-
-public class Insurance {
-    private int insuranceIdx;
-    private String content;
-    private int fee;
-    private int maxReward;
-    private int period;
-    private int rate;
-    private int employeeIdx;
-    public Contract contract;
-    public Accident accident;
-
-
-}
