+++ server/src/main/java/domain/User.java
@@ -1,8 +1,25 @@
 package domain;
 
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@Setter
 public class User {
+
+    @Id
+    @GeneratedValue
     private int userIdx;
-//    private String address;
+
+    @Embedded
+    private Address address;
+
     private boolean checkSub;
     private boolean checkSug;
     private String email;
@@ -16,7 +33,5 @@ public class User {
     public Contract contract;
     public Accident accident;
 
-
-    private Address address;
 
 }
