+++ server/src/main/java/domain/ex_insurance/Travel.java
@@ -1,9 +1,22 @@
 package domain.ex_insurance;
 
-import domain.Insurance;
+import lombok.Getter;
+import lombok.Setter;
 
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@DiscriminatorValue("D")
+@Getter
+@Setter
 public class Travel extends Insurance {
+    @Id
+    @GeneratedValue
     private int travelIdx;
+
     private int country;
     private int period;
